# Fix order of do-while loop in transpile optimisation